% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itpde-package.R
\docType{package}
\name{itpde-package}
\alias{itpde}
\alias{itpde-package}
\title{itpde: The International Trade and Production Database for Estimation (ITPD-E)}
\description{
Provides consistent data on international and domestic trade for 243 countries, 170 industries, and 17 years. The data are constructed at the industry level covering agriculture, mining, energy, manufacturing, and services for the period 1986-2019. The ITPD-E is constructed using reported administrative data and intentionally does not include information estimated by statistical techniques, which makes the ITPD-E well suited for estimation of economic models, such as the gravity model of trade.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/itpde}
  \item Report bugs at \url{https://github.com/pachadotdev/itpde/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{hello+r@pacha.dev} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item  United States International Trade Comission (USITC) [data contributor]
}

}
\keyword{internal}
