% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{itpde_connect}
\alias{itpde_connect}
\title{Connect to the itpde database}
\usage{
itpde_connect(dir = itpde_path())
}
\arguments{
\item{dir}{Database location on disk. Defaults to the \code{itpde}
directory inside the R user folder or the \code{ITPDE_DIR} environment variable
if specified.}
}
\description{
Returns a local database connection. This corresponds to a DBI-compatible
DuckDB database.
}
\examples{
\dontrun{
 DBI::dbListTables(itpde_connect())

 DBI::dbGetQuery(
  itpde_connect(),
  'SELECT * FROM itpde WHERE year = 2010'
 )
}
}
