% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{itpde_delete}
\alias{itpde_delete}
\title{Delete the itpde database from your computer}
\usage{
itpde_delete(ask = TRUE)
}
\arguments{
\item{ask}{If so, a menu will be displayed to confirm the action to
delete any existing itpde database. By default it is true.}
}
\description{
Deletes the \code{itpde} directory and all of its contents, including
all versions of the itpde database created with any DuckDB version.
}
\examples{
\dontrun{ itpde_delete() }
}
