% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{itpde_disconnect}
\alias{itpde_disconnect}
\title{Disconnect the itpde database}
\usage{
itpde_disconnect()
}
\description{
An auxiliary function to disconnect from the database.
}
\examples{
itpde_disconnect()
}
