% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{itpde_download}
\alias{itpde_download}
\title{Download the ITPD-E Database to your computer}
\usage{
itpde_download(ver = NULL)
}
\arguments{
\item{ver}{The version to download. By default it is the latest version
available on GitHub. All versions can be viewed at
\url{https://github.com/pachadotdev/itpde/releases}.}
}
\description{
This command downloads the entire database as a single zip file that is
unzipped to create the local database. If you do not want to download the
database on your home computer, run usethis::edit_r_environ() to create the
environment variable ITPDE_DIR with the path.
}
\examples{
\dontrun{ itpde_download() }
}
