% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itpde-package.R
\docType{data}
\name{region_names}
\alias{region_names}
\title{Region names: Auxiliary table}
\format{
A data frame with 15 rows and 2 columns:
|Variable name         |Variable description                                                                                                |
|region_id             |Region code                                                                                                         |
|region                |Region name                                                                                                         |
}
\source{
Adapted from the original DGD dataset
}
\description{
Region names: Auxiliary table
}
\author{
Own creation, adapted from the original DGD dataset
}
\keyword{data}
