% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itpde-package.R
\docType{package}
\name{usitcgravity-package}
\alias{usitcgravity}
\alias{usitcgravity-package}
\title{usitcgravity: USITC Gravity Database Adapted From the International Trade and Production Database for Estimation (ITPD-E) and Dynamic Gravity Dataset (DGD)}
\description{
Provides consistent data on international and domestic trade for 243 countries, 170 industries, and 17 years alongside variables for gravity estimation. The data sources are ITPD-E and DGD from the USITC, which use administrative data and intentionally does not include information estimated by statistical techniques, which makes the datasets suitable for gravity estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/usitcgravity}
  \item Report bugs at \url{https://github.com/pachadotdev/usitcgravity/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{hello+r@pacha.dev} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item  United States International Trade Comission (USITC) [data contributor]
}

}
\keyword{internal}
