% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{usitcgravity_connect}
\alias{usitcgravity_connect}
\title{Connect to the usitcgravity database}
\usage{
usitcgravity_connect(dir = usitcgravity_path())
}
\arguments{
\item{dir}{Database location on disk. Defaults to the \code{usitcgravity}
directory inside the R user folder or the \code{USITCGRAVITY_DIR} environment variable
if specified.}
}
\description{
Returns a local database connection. This corresponds to a DBI-compatible
DuckDB database.
}
\examples{
\dontrun{
 DBI::dbListTables(usitcgravity_connect())

 DBI::dbGetQuery(
  usitcgravity_connect(),
  'SELECT * FROM usitcgravity WHERE year = 2010'
 )
}
}
