% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{usitcgravity_delete}
\alias{usitcgravity_delete}
\title{Delete the usitcgravity database from your computer}
\usage{
usitcgravity_delete(ask = TRUE)
}
\arguments{
\item{ask}{If so, a menu will be displayed to confirm the action to
delete any existing usitcgravity database. By default it is true.}
}
\description{
Deletes the \code{usitcgravity} directory and all of its contents, including
all versions of the usitcgravity database created with any DuckDB version.
}
\examples{
\dontrun{ usitcgravity_delete() }
}
