% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{usitcgravity_disconnect}
\alias{usitcgravity_disconnect}
\title{Disconnect the usitcgravity database}
\usage{
usitcgravity_disconnect()
}
\description{
An auxiliary function to disconnect from the database.
}
\examples{
usitcgravity_disconnect()
}
